requirements.txt
numpy
pillow
flask
flask_cors